Change DAO query parameter type

+++ app/src/main/java/com/example/ntu_timetable_calendar/DAO/TaskDAO.java
@@ -36,5 +36,5 @@ public interface TaskDAO {
     LiveData<TaskEntity> getTask(int id);
 
     @Query("SELECT * FROM task_table WHERE deadLine BETWEEN :nowTime AND :deadLineTime")
-    LiveData<List<TaskEntity>> getTasksWithinTime(int nowTime ,long deadLineTime);
+    LiveData<List<TaskEntity>> getTasksWithinTime(long nowTime ,long deadLineTime);
 }
